♻️ extract browserChecks
